fix: use kotlin 1.6.20 because of MABS 9
Context: Because of MABS 9 builds, we need to compile the library with Kotlin 1.6.20 and not 1.9.10. If we only needed to support MABS 10, we could use 1.9.10. Because of this, we needed to downgrade some our the gradle dependencies.

One of these dependencies is 'androidx.camera:camera-core', which we had to downgrade. For that reason, we cannot use method ImageProxy.toBitmap any longer and had to implement our own version of it.

References: https://outsystemsrd.atlassian.net/browse/RMET-2894