feat: implement UI for tablets (portrait and landscape)
Context: According to Android's documentation, we should use the Window size classes to determine if the device the app is running on is a phone or a tablet. When in Portrait, we should use the window width to determine this (99.96% of phones will have a compact width). When in Landscape, the height of the window should be used (99.78% of phones will have a compact height). More info here: https://developer.android.com/guide/topics/large-screens/support-different-screen-sizes#window_size_classes, and here: https://developer.android.com/jetpack/compose/layouts/adaptive

Why we are using the landscape UI (ScanScreenUILandscape) for tablets for both orientations, portrait and landscape: because for tablets, the arrangement of the elements in the screen is always the same, in a row, where the buttons are to the right of the scan area (rectangle). This way, we can avoid code repetition.

References: https://outsystemsrd.atlassian.net/browse/RMET-2912