feat: refactor code for barcode analysis
Context: Since we want to "crop" the image passed to analyze, we need to use a Bitmap object to represent the image, so that we can "crop" it. This wouldn't be possible using the ImageProxy object directly. Since both libraries (ZXing and MLKit) now use a Bitmap, we might as well refactor the code and avoid code repetition.

References: https://outsystemsrd.atlassian.net/browse/RMET-2912