refactor: use coroutines to work with bitmaps
Context: In Android, when working with bitmaps, which is a type of image processing, it is recommended to use coroutines so that the main UI thread isn't blocked.

Dispatchers.Default is used because it is optimized for tasks that are CPU intensive, like image processing and transformation, since it provides a limited number of threads that is equal to the number of CPU cores.

More info: https://developer.android.com/develop/ui/views/graphics and https://developer.android.com/kotlin/coroutines/coroutines-adv