refactor: run scanCode inside coroutine
Context: In Android, when working with bitmaps, which is a type of image processing, it is recommended to use coroutines so that the main UI thread isn't blocked.

We are putting the whole call to the scanCode feature inside a coroutine because we want to run the whole scanning feature inside a single thread. If we put just the code that works with bitmaps inside a coroutine, we would be launching a thread for each frame that is sent for analysis, which could potentially lead to memory issues.

Dispatchers.Default is used because it is optimized for tasks that are CPU intensive, like image processing and transformation, since it provides a limited number of threads that is equal to the number of CPU cores.

More info: https://developer.android.com/develop/ui/views/graphics and https://developer.android.com/kotlin/coroutines/coroutines-adv