feat: set fallback rule for resolution
Context: We're requesting a resolution of 1920x1080 for the frames that will be scanned. If this resolution isn't available on the device, we set the fallback to the closest highest resolution. If that also isn't available, then it will fallback to the closest lower resolution.

References: https://outsystemsrd.atlassian.net/browse/RMET-3399